Build the gRPC Java library from source
Build the library from source since the imported prebuilts are
inconsistent and incomplete. For starters, the source is that of version
1.15.0 but the imported jars are those of version 1.14.0.

More importantly the grpc-java-protobuf library depends on the
grpc-java-protobuf-lite library that is not prebuilt. This causes
runtime errors due to missing symbols that should have been caught at
compile-time.

The Soong build modules follow the upstream Bazel rules as closely as
possible. Note that some static libraries were added to the grpc-java to
maintain compatibility with other rules that depend on it since the
original prebuilt grpc-java-core jar contained all sources under core/
which is not how the library is built using Bazel.

Test: m grpc-java
Test: Diffed the entries between the old and new jars
Bug: 148404241
Change-Id: I060333a68848272d5d9e09c040c1e35451a408c6