Remove gRPC Netty shaded static library
Remove the build module for the shaded library which causes issues at
runtime due to missing symbols. This happens because the jar is prebuilt
from an older version of gRPC and is no longer compatible with the
current source.

Additionally, the grpc-java module should not statically link against a
specific transport jar that contains service providers. Doing so forces
an automatically-loaded implementation on users.

Test: m grpc-java
Bug: 148404241
Change-Id: I77589c05756b55146648d726ea6e235f1887ac38