Add a build module for the Netty Shaded library
The gRPC documentation recommends that most users use the Netty Shaded
library for Transport Layer Security (TLS) on non-Android platforms. The
library prepackages netty-tcnative on BoringSSL which makes it easier to
use.

Note that this change essentially restores the module which was deleted
in a previous commit. This new binary is imported for the same version
that the source is currently at and does not cause any runtime issues.

Test: m grpc-java-netty-shaded
Bug: 148404241
Change-Id: I7ae611add53627d2c00beeeac829fa31069d5594