Add a lite OkHttp gRPC Client Library.
Updates several targets within grpc-grpc-java to allow for Android and Host targets to allow the building of a gRPC client library that uess OkHttp and Protobuf Lite. This provides a lightweight client library for gRPC clients that works on both Android and Host since it uses lite protos rather than full protos.

Test: m grpc-java-okhttp-client-lite
Test: mm
Test: Manual testing using client library on Android.
Change-Id: Ib9dc14a44856914b38662c01905337f98bd8d108